% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataElement_table.R
\name{get_dataElement_table}
\alias{get_dataElement_table}
\title{Get dataElement table with dataElement group names}
\usage{
get_dataElement_table(username, password, base_url)
}
\arguments{
\item{username}{user credentials}

\item{password}{user credentials}

\item{base_url}{base url of the DHIS2 instance}
}
\value{
a tibble of data elements with data element group names
}
\description{
Get dataElement table with dataElement group names
}
\examples{
\dontrun{
  get_dataElement_table(username, password, base_url)
}

}
