% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataElementGroup.R
\name{get_dataElementGroup}
\alias{get_dataElementGroup}
\title{Get data element groups from DHIS2}
\usage{
get_dataElementGroup(username, password, base_url)
}
\arguments{
\item{username}{username credentials}

\item{password}{password credentials}

\item{base_url}{base url of the DHIS2 instance}
}
\value{
a list of data element groups
}
\description{
Get data element groups from DHIS2
}
\examples{
\dontrun{
  get_dataElementGroup(username, password, base_url)
}

}
