% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataElement_table.R
\name{get_dataElements}
\alias{get_dataElements}
\title{Extract data elements from DHIS2}
\usage{
get_dataElements(username, password, base_url)
}
\arguments{
\item{username}{user credentials}

\item{password}{user credentials}

\item{base_url}{base url of the DHIS2 instance}
}
\value{
a tibble of data elements
}
\description{
Extract data elements from DHIS2
}
\examples{
\dontrun{
  get_dataElements(username, password, base_url)
}

}
