% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator.R
\name{get_indicatorGroup}
\alias{get_indicatorGroup}
\title{Fetch indicator groups from DHIS2}
\usage{
get_indicatorGroup(username, password, base_url)
}
\arguments{
\item{username}{username}

\item{password}{password}

\item{base_url}{base_url from DHIS2 instance}
}
\value{
A tibble of indicator groups from DHIS2
}
\description{
Fetch indicator groups from DHIS2
}
\examples{
\dontrun{
  get_indicatorGroup(username, password, base_url)
}

}
