% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_table.R
\name{get_indicators}
\alias{get_indicators}
\title{Pull Indicators from DHIS2}
\usage{
get_indicators(username, password, base_url)
}
\arguments{
\item{username}{credentials for DHIS2}

\item{password}{password credentials for DHIS2}

\item{base_url}{base url for DHIS2 instance}
}
\value{
A tibble of indicators from DHIS2
}
\description{
Pull Indicators from DHIS2
}
\examples{
\dontrun{
  get_indicators(username, password, base_url)
}

}
