% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_table.R
\name{get_indicators_table}
\alias{get_indicators_table}
\title{Create extended indicators table including indicator group name}
\usage{
get_indicators_table(username, password, base_url)
}
\arguments{
\item{username}{user credentials}

\item{password}{user credentials}

\item{base_url}{base_url for DHIS2 instance}
}
\value{
a tibble of indicators with indicator group
}
\description{
Create extended indicators table including indicator group name
}
\examples{
\dontrun{
  get_indicators_table(username, password, base_url)
}

}
