% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_table.R
\name{get_indicatorType}
\alias{get_indicatorType}
\title{Table of indicator types from DHIS2}
\usage{
get_indicatorType(username, password, base_url)
}
\arguments{
\item{username}{username}

\item{password}{password}

\item{base_url}{base_url from DHIS2 instance}
}
\value{
a tibble of indicatortypes
}
\description{
Table of indicator types from DHIS2
}
\examples{
\dontrun{
  get_indicatorType(username, password, base_url)
}

}
