% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_ou_hierarchy}
\alias{get_ou_hierarchy}
\title{Extract OU hierarchy from OU table}
\usage{
get_ou_hierarchy(ou_table)
}
\arguments{
\item{ou_table}{Organisation Unit Table}
}
\value{
clean hierarhcy
}
\description{
Extract OU hierarchy from OU table
}
\examples{
\dontrun{
  get_ou_hierarchy(ou_table)
}

}
