% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outable.R
\name{get_outable}
\alias{get_outable}
\title{Create a table for Organisation Unit.}
\usage{
get_outable(username, password, base_url)
}
\arguments{
\item{username}{credentials}

\item{password}{credentials}

\item{base_url}{base_url of DHIS2 instances}
}
\value{
a tibble of OUs
}
\description{
Create a table for Organisation Unit.
}
\examples{
\dontrun{
  get_outable(username, password, base_url)
}

}
