% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_table.R
\name{get_programIndicator}
\alias{get_programIndicator}
\title{Get all program inidcators}
\usage{
get_programIndicator(username, password, base_url)
}
\arguments{
\item{username}{DHIS2 username}

\item{password}{DHIS2 password}

\item{base_url}{base_url for DHIS2 instance}
}
\value{
a tibble of program inidcators
}
\description{
Get all program inidcators
}
\examples{
\dontrun{
  get_programIndicator(username, password, base_url)
}

}
