% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_programIndicator_table.R
\name{get_programIndicator_table}
\alias{get_programIndicator_table}
\title{Program Indicator Table}
\usage{
get_programIndicator_table(username, password, base_url)
}
\arguments{
\item{username}{DHIS2 username}

\item{password}{DHIS2 password}

\item{base_url}{DHIS2 base_url}
}
\value{
a tibble of program indicators
}
\description{
Program Indicator Table
}
\examples{
\dontrun{
  get_programIndicator_table(username, password, base_url)
}

}
