% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicator_table.R
\name{get_programIndicatorGroup}
\alias{get_programIndicatorGroup}
\title{Get a tibble of program indicators groups}
\usage{
get_programIndicatorGroup(username, password, base_url)
}
\arguments{
\item{username}{DHIS2 username}

\item{password}{DHIS2 password}

\item{base_url}{base_url for DHIS2 instance}
}
\value{
a tibble of program inidcator groups
}
\description{
Get a tibble of program indicators groups
}
\examples{
\dontrun{
  get_programIndicatorGroup(username, password, base_url)
}

}
