% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_programIndicator_table.R
\name{get_programs}
\alias{get_programs}
\title{Get a tibble of all programs}
\usage{
get_programs(username, password, base_url)
}
\arguments{
\item{username}{DHIS2 username}

\item{password}{DHIS2 password}

\item{base_url}{base_url for DHIS2 instance}
}
\value{
a tibble of programs
}
\description{
Get a tibble of all programs
}
\examples{
\dontrun{
  get_programs(username, password, base_url)
}

}
