% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dhis2_element.R
\name{pull_dhis2_element}
\alias{pull_dhis2_element}
\title{Pull out an element list from DHIS2}
\usage{
pull_dhis2_element(base_url, username, password)
}
\arguments{
\item{base_url}{base_url of DHIS2 instance}

\item{username}{username for DHIS2 instance}

\item{password}{password for DHIS2 instance}
}
\value{
A list of elements from DHIS2
}
\description{
Pull out an element list from DHIS2
}
\examples{
\dontrun{
  pull_dhis2_element(base_url, username, password)
}

}
