% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ou_table.R
\name{get_ou_table}
\alias{get_ou_table}
\title{Extract OU hierarchy from OU table}
\usage{
get_ou_table(username, password, url)
}
\arguments{
\item{username}{credentials}

\item{password}{credentials}

\item{url}{base_url of DHIS2 instance}
}
\value{
Org Unit Table with different levels
}
\description{
Extract OU hierarchy from OU table
}
\examples{
\dontrun{
  get_ou_table(username, password,url)
}

}
